requirements.txt
django 
djangorestframework 
djangorestframework-simplejwt
django-admin-tailwind
python-decouple
pillow
Faker